Better usage instructions for traject 3.0+
Where this will no longer be a default and an automatic dependency.

Note that the docs saying marc4j_reader.permissive defaulted to true were kind of wrong,
it defaulted to false/nil in this code, but the traject core gem previously made it default to true,
which it no longer does.